Fix INTEGER column value overflow.

+++ src/android-native/src/main/java/io/liteglue/SQLiteAndroidDatabase.java
@@ -335,7 +335,7 @@ class SQLiteAndroidDatabase {
                 row.putNull(key);
                 break;
             case Cursor.FIELD_TYPE_INTEGER:
-                row.putInt(key, cur.getInt(i));
+                row.putDouble(key, cur.getLong(i));
                 break;
             case Cursor.FIELD_TYPE_FLOAT:
                 row.putDouble(key, cur.getDouble(i));
